requirements.txt
python_pptx==0.6.18
autocorrect==1.1.0
nltk==3.5
